% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_attrsenti.R
\name{attribute_sentiment_classifer}
\alias{attribute_sentiment_classifer}
\title{attribute_sentiment_classifer() function
Gives attribute class and sentiment class for a sentence
Can give both binary and granular sentiment based on type argument}
\usage{
attribute_sentiment_classifer(sentence, type)
}
\arguments{
\item{type, }{type=1 is binary and type=2 is granular}
}
\description{
attribute_sentiment_classifer() function
Gives attribute class and sentiment class for a sentence
Can give both binary and granular sentiment based on type argument
}
