% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_attr.R
\name{getattr}
\alias{getattr}
\title{getattr() function that prints the attribute class of a word}
\usage{
getattr(word)
}
\arguments{
\item{word}{Give a word whose attribute you want to classify}
}
\description{
getattr() function that prints the attribute class of a word
}
