% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_senti.R
\name{getbinsentiment}
\alias{getbinsentiment}
\title{getbinsentiment() function to print the binary sentiment associated with a word}
\usage{
getbinsentiment(word)
}
\description{
getbinsentiment() function to print the binary sentiment associated with a word
}
