% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_senti.R
\name{getgransentiment}
\alias{getgransentiment}
\title{getgransentiment() function to print the granular sentiment associated with a word}
\usage{
getgransentiment(word)
}
\description{
getgransentiment() function to print the granular sentiment associated with a word
}
