% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_attr.R
\name{isattr}
\alias{isattr}
\title{isattr() function tells whether word is an attribute word}
\usage{
isattr(word)
}
\arguments{
\item{word}{Give a word whose attribute you want to classify}
}
\value{
boolean True or False, True if word is attribute word
}
\description{
isattr() function tells whether word is an attribute word
}
