% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_attr.R
\name{retattr}
\alias{retattr}
\title{retattr() function returns the attribute class of an attribute word}
\usage{
retattr(word)
}
\arguments{
\item{word}{Give a word whose attribute you want to classify
#'@return string , attribute word}
}
\description{
retattr() function returns the attribute class of an attribute word
}
