% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_senti.R
\name{retsentiment}
\alias{retsentiment}
\title{sentiment() function
This function allows you to get binary (Positive,Neutral,Negative)classes
more granular(1,2,3,4,5) classes for sentiments
for frequent sentiment words appearing in restaurant reviews
for binary pass type as 1 and for granular pass 2}
\usage{
retsentiment(word, type)
}
\description{
sentiment() function
This function allows you to get binary (Positive,Neutral,Negative)classes
more granular(1,2,3,4,5) classes for sentiments
for frequent sentiment words appearing in restaurant reviews
for binary pass type as 1 and for granular pass 2
}
